requirements.txt
anyio==3.6.1
ariadne==0.16.0
autopep8==1.7.0
black==22.8.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
essential-generators==1.0
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
graphql-core==3.2.1
greenlet==1.1.3
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
names==0.3.0
numpy==1.23.3
pandas==1.5.0
pathspec==0.10.1
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
yapf==0.32.0
